requirements.txt
asgiref==3.7.2
black==23.10.1
click==8.1.7
Django==4.1.7
django-stubs==4.2.6
django-stubs-ext==4.2.5
mypy==1.6.1
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
protobuf==4.21.12
python-json-logger==2.0.7
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.7.1
